feat(components): migrate Input to TypeScript
