+++ src/components/Input/Input.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import Input from './Input';
+import { Input } from './Input';
 
 # Input
 
@@ -9,7 +9,6 @@ A text field is an input field for inserting short text and numbers. It has
 multiple functions and supports multiple character types/formats.
 
 <Story id="forms-input--base" />
-
 <Props of={Input} />
 
 ## When to use it
