+++ src/components/Input/Input.spec.tsx
@@ -16,9 +16,13 @@
 import React from 'react';
 import { css } from '@emotion/core';
 
+import { create, render, renderToHtml, axe } from '../../util/test-utils';
+
 import Input from '.';
 
-const DummyElement = () => <div style={{ width: '24px', height: '24px' }} />;
+const DummyElement = (props: { className?: string }) => (
+  <div style={{ width: '24px', height: '24px' }} {...props} />
+);
 
 describe('Input', () => {
   /**
@@ -126,7 +130,7 @@ describe('Input', () => {
      * Should accept a working ref for input
      */
     it('should accept a working ref', () => {
-      const tref = React.createRef();
+      const tref = React.createRef<HTMLInputElement & HTMLTextAreaElement>();
       const { container } = render(<Input ref={tref} />);
       const input = container.querySelector('input');
       expect(tref.current).toBe(input);
@@ -136,7 +140,7 @@ describe('Input', () => {
      * Should accept a working ref for textarea
      */
     it('should accept a working ref also for textarea', () => {
-      const tref = React.createRef();
+      const tref = React.createRef<HTMLInputElement & HTMLTextAreaElement>();
       const { container } = render(<Input as="textarea" ref={tref} />);
       const textarea = container.querySelector('textarea');
       expect(tref.current).toBe(textarea);
